chore: move session code into its own module
